requirements.txt
Django==4.2.7
python-docx==0.8.11
PyPDF2==3.0.1
reportlab==4.0.4
openpyxl==3.1.2
django-cleanup==8.0.0
python-magic==0.4.27
django-widget-tweaks==1.5.0
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
Pillow>=11.0.0
pdf2image==1.16.3
dj-database-url
# Security Packages
python-magic-bin==0.4.24  # File type validation
django-axes==6.0.1  # Login attempt limiting
django-cors-headers==4.2.0  # CORS protection
django-ipware==5.0.0  # IP address detection
python-decouple==3.8  # Environment variables
django-redis==5.3.0  # Redis cache
redis==5.0.1  # Redis client
django-environ==0.11.2  # Environment configuration
argon2-cffi==23.1.0  # Secure password hashing
psycopg2-binary==2.9.9  # PostgreSQL for production
whitenoise==6.5.0  # Static files with security
django-csp==3.8  # Content Security Policy
sentry-sdk==1.35.0  # Error tracking
django-ratelimit==4.0.0  # Rate limiting
django-auditlog==2.3.0  # Audit logging

